Arm collision
Using the so called "unga-bunga" method coined by Rowinski. This method uses a rectangle from the center of the player to the tip of the arm for collision.

Together with this is a raycast sent from the tip of the arm to the mouse which snaps the arm's solver to the edge of the collided object.

Scaled everything up by 5.

Disabled rendering of the 3D projection when the game isn't started (in the editor)