Navmesh navigation and combat
The enemy can now fully utilize a navmesh to navigate to any given position (inside of the navmesh). Both the navmesh and current path are automatically updated at an interval to allow for real-time navigation.

The first part of the combat system is made. When the enemy is in combat mode it will scan the area and select a strategic position to move to before attacking the player. The parameters for this can be customized.

The arm has basic movement now and will attempt to aim towards the player.

PID controller and settings have been greatly improved.